types

+++ src/core/test_core_api.c
@@ -125,7 +125,8 @@ disconnect_notify (void *cls,
 static unsigned int
 bfc_callback (void *cls,
               const struct GNUNET_PeerIdentity *receiver,
-              void *position, unsigned int padding)
+              void *position, 
+	      size_t padding)
 {
   GNUNET_log (GNUNET_ERROR_TYPE_DEBUG,
               "Core requests data for `%4s', I have none.\n",
